requirements.txt
asgiref==3.5.2
Django==4.1.1
django-appconf==1.0.5
django-ckeditor==6.5.1
django-compressor==4.1
django-crispy-forms==1.14.0
django-filter==22.1
django-js-asset==2.0.0
django-libsass==0.9
gunicorn==20.1.0
libsass==0.21.0
Pillow==9.2.0
rcssmin==1.1.0
rjsmin==1.2.0
six==1.16.0
sqlparse==0.4.2
tzdata==2022.2
